refactor(spark-core): change names for storage classes
