+++ packages/spark-core/src/lib/storage/index.js
@@ -3,8 +3,8 @@ export {
   waitForValue,
   waitForValueAndPersistResult
 } from './decorators';
-export {default as Storage} from './storage';
-export {default as default} from './storage';
-export {default as PluginStorage} from './plugin-storage';
-export {default as StorageAdapter} from './storage-adapter';
+export {default as Store} from './store';
+export {default as default} from './store';
+export {default as SparkPluginStorage} from './spark-plugin-storage';
+export {default as MemoryStoreAdapter} from './memory-store-adapter';
 export {StorageError, NotFoundError} from './errors';
