+++ packages/spark-core/src/lib/storage/memory-store-adapter.js
